Many commands removed, added Server Info button, removed lv_LV.lang

+++ src/main/java/latmod/ftbu/mod/client/gui/GuiViewImage.java
@@ -44,8 +44,8 @@ public class GuiViewImage extends GuiLM
 		{
 			setTexture(texCoords.texture);
 			
-			int w = texCoords.width;
-			int h = texCoords.height;
+			double w = texCoords.width;
+			double h = texCoords.height;
 			
 			if(w > width)
 			{
